+++ src/consumer.ts
@@ -1,12 +1,19 @@
 import { SQSEvent } from 'aws-lambda';
-import { Resource } from './resource/Resource';
-import ResourceEntity from './db/model/Resource';
+import ResourceEntity from './db/resource/model/Resource';
+import Resource from './db/resource/model/Resource';
+import sequelize from './db/connection';
+
+sequelize.addModels([ResourceEntity]);
+ResourceEntity.sync();
 
 export const handler = async (event: SQSEvent): Promise<void> => {
   for (const record of event.Records) {
     const body = JSON.parse(record.body) as Resource;
 
-    console.log('Received message: ', body);
+    const resource = await ResourceEntity.create({body})
 
+    await resource.save();
+
+    console.log('Received message: ', body);
   }
 };
