+++ src/db/connection.ts
@@ -1,18 +1,17 @@
 
 import { Sequelize } from "sequelize-typescript";
-import ResourceEntity from "./model/Resource";
 
-const sequelize = new Sequelize('database_name', 'username', 'password', {
-    host: '', 
+const sequelize = new Sequelize(process.env.DB_NAME, process.env.DB_USER, process.env.DB_PASSWORD, {
+    host: process.env.DB_HOST,
     port: 3306,
-    dialect: 'mysql'
-});
-
-sequelize.addModels([ResourceEntity])
-sequelize.sync().then(() => {
-    console.log('Resource table created successfully');
-}).catch((error) => {
-    console.error('Error creating User table', error);
+    dialect: 'mysql',
+    pool: {
+        max: 2,
+        min: 0,
+        idle: 0,
+        acquire: 3000,
+        evict: 30000
+      }
 });
 
 export default sequelize;
